src: refactor project gym pybullet
