+++ gym_pybullet_drones/gym_pybullet_drones/assets/architrave copy.urdf
@@ -1,0 +1,33 @@
+<?xml version="1.0" ?>
+
+<robot name="architrave">
+
+  <link name="base_link"> 
+
+    <inertial>
+      <origin rpy="0 0 0" xyz="0 0 0"/>
+      <mass value="0.3"/>
+      <inertia ixx="0" ixy="0.0" ixz="0.0" iyy="0" iyz="0.0" izz="0"/>
+    </inertial>
+
+    <visual>
+      <origin rpy="0 0 0" xyz="0 0 0"/>
+      <geometry>
+        <box size=".8 .05 .05" />
+      </geometry> 
+      <material name="grey">
+        <color rgba=".5 .5 .5 1"/>
+      </material> 
+    </visual>
+
+    <collision>
+      <origin rpy="0 0 0" xyz="0 0 0"/>
+      <geometry>
+        <box size=".8 .05 .05" />
+      </geometry>
+    </collision>  
+
+  </link>
+
+</robot>
+
