+++ gym_pybullet_drones/gym_pybullet_drones/assets/box.urdf
@@ -1,0 +1,16 @@
+<?xml version="1.0" ?>
+<robot name="test.urdf">
+  <link name="baseLink">
+    <inertial>
+      <origin rpy="0 0 0" xyz="0 0 0"/>
+       <mass value=".1"/>
+       <inertia ixx="1" ixy="0" ixz="0" iyy="1" iyz="0" izz="1"/>
+    </inertial>
+    <visual>
+      <origin rpy="0 0 0" xyz="0 0 0"/>
+      <geometry>
+        <box size=".05 .05 .1" />
+      </geometry>
+    </visual>
+  </link>
+</robot>
