+++ gym_pybullet_drones/gym_pybullet_drones/control/DSLPIDControl.py
@@ -45,9 +45,19 @@ class DSLPIDControl(BaseControl):
         self.MIN_PWM = 20000
         self.MAX_PWM = 65535
         if self.DRONE_MODEL == DroneModel.CF2X:
-            self.MIXER_MATRIX = np.array([ [.5, -.5,  -1], [.5, .5, 1], [-.5,  .5,  -1], [-.5, -.5, 1] ])
+            self.MIXER_MATRIX = np.array([ 
+                                    [-.5, -.5, -1],
+                                    [-.5,  .5,  1],
+                                    [.5, .5, -1],
+                                    [.5, -.5,  1]
+                                    ])
         elif self.DRONE_MODEL == DroneModel.CF2P:
-            self.MIXER_MATRIX = np.array([ [0, -1,  -1], [+1, 0, 1], [0,  1,  -1], [-1, 0, 1] ])
+            self.MIXER_MATRIX = np.array([
+                                    [0, -1,  -1],
+                                    [+1, 0, 1],
+                                    [0,  1,  -1],
+                                    [-1, 0, 1]
+                                    ])
         self.reset()
 
     ################################################################################
