+++ gym_pybullet_drones/gym_pybullet_drones/examples/downwash.py
@@ -26,7 +26,6 @@ DEFAULT_GUI = True
 DEFAULT_RECORD_VIDEO = False
 DEFAULT_SIMULATION_FREQ_HZ = 240
 DEFAULT_CONTROL_FREQ_HZ = 48
-DEFAULT_AGGREGATE = True
 DEFAULT_DURATION_SEC = 12
 DEFAULT_OUTPUT_FOLDER = 'results'
 DEFAULT_COLAB = False
@@ -37,7 +36,6 @@ def run(
         record_video=DEFAULT_RECORD_VIDEO, 
         simulation_freq_hz=DEFAULT_SIMULATION_FREQ_HZ, 
         control_freq_hz=DEFAULT_CONTROL_FREQ_HZ, 
-        aggregate=DEFAULT_AGGREGATE, 
         duration_sec=DEFAULT_DURATION_SEC,
         output_folder=DEFAULT_OUTPUT_FOLDER,
         plot=True,
@@ -45,14 +43,13 @@ def run(
     ):
     #### Initialize the simulation #############################
     INIT_XYZS = np.array([[.5, 0, 1],[-.5, 0, .5]])
-    AGGR_PHY_STEPS = int(simulation_freq_hz/control_freq_hz) if aggregate else 1
     env = CtrlAviary(drone_model=drone,
                      num_drones=2,
                      initial_xyzs=INIT_XYZS,
                      physics=Physics.PYB_DW,
                      neighbourhood_radius=10,
-                     freq=simulation_freq_hz,
-                     aggregate_phy_steps=AGGR_PHY_STEPS,
+                     pyb_freq=simulation_freq_hz,
+                     ctrl_freq=control_freq_hz,
                      gui=gui,
                      record=record_video,
                      obstacles=True
@@ -67,7 +64,7 @@ def run(
     wp_counters = np.array([0, int(NUM_WP/2)])
 
     #### Initialize the logger #################################
-    logger = Logger(logging_freq_hz=int(simulation_freq_hz/AGGR_PHY_STEPS),
+    logger = Logger(logging_freq_hz=control_freq_hz,
                     num_drones=2,
                     duration_sec=duration_sec,
                     output_folder=output_folder,
@@ -78,43 +75,38 @@ def run(
     ctrl = [DSLPIDControl(drone_model=drone) for i in range(2)]
 
     #### Run the simulation ####################################
-    CTRL_EVERY_N_STEPS = int(np.floor(env.SIM_FREQ/control_freq_hz))
-    action = {str(i): np.array([0, 0, 0, 0]) for i in range(2)}
+    action = np.zeros((2,4))
     START = time.time()
-    for i in range(0, int(duration_sec*env.SIM_FREQ), AGGR_PHY_STEPS):
+    for i in range(0, int(duration_sec*env.CTRL_FREQ)):
 
         #### Step the simulation ###################################
-        obs, reward, done, info = env.step(action)
+        obs, reward, terminated, truncated, info = env.step(action)
 
-        #### Compute control at the desired frequency ##############
-        if i%CTRL_EVERY_N_STEPS == 0:
+        #### Compute control for the current way point #############
+        for j in range(2):
+            action[j, :], _, _ = ctrl[j].computeControlFromState(control_timestep=env.CTRL_TIMESTEP,
+                                                                    state=obs[j],
+                                                                    target_pos=np.hstack([TARGET_POS[wp_counters[j], :], INIT_XYZS[j, 2]]),
+                                                                    )
 
-            #### Compute control for the current way point #############
-            for j in range(2):
-                action[str(j)], _, _ = ctrl[j].computeControlFromState(control_timestep=CTRL_EVERY_N_STEPS*env.TIMESTEP,
-                                                                       state=obs[str(j)]["state"],
-                                                                       target_pos=np.hstack([TARGET_POS[wp_counters[j], :], INIT_XYZS[j, 2]]),
-                                                                       )
-
-            #### Go to the next way point and loop #####################
-            for j in range(2):
-                wp_counters[j] = wp_counters[j] + 1 if wp_counters[j] < (NUM_WP-1) else 0
+        #### Go to the next way point and loop #####################
+        for j in range(2):
+            wp_counters[j] = wp_counters[j] + 1 if wp_counters[j] < (NUM_WP-1) else 0
 
         #### Log the simulation ####################################
         for j in range(2):
             logger.log(drone=j,
-                       timestamp=i/env.SIM_FREQ,
-                       state=obs[str(j)]["state"],
+                       timestamp=i/env.CTRL_FREQ,
+                       state=obs[j],
                        control=np.hstack([TARGET_POS[wp_counters[j], :], INIT_XYZS[j ,2], np.zeros(9)])
                        )
 
         #### Printout ##############################################
-        if i%env.SIM_FREQ == 0:
-            env.render()
+        env.render()
 
         #### Sync the simulation ###################################
         if gui:
-            sync(i, START, env.TIMESTEP)
+            sync(i, START, env.CTRL_TIMESTEP)
 
     #### Close the environment #################################
     env.close()
@@ -136,7 +128,6 @@ if __name__ == "__main__":
     parser.add_argument('--record_video',       default=DEFAULT_RECORD_VIDEO,      type=str2bool,      help='Whether to record a video (default: False)', metavar='')
     parser.add_argument('--simulation_freq_hz', default=DEFAULT_SIMULATION_FREQ_HZ,        type=int,           help='Simulation frequency in Hz (default: 240)', metavar='')
     parser.add_argument('--control_freq_hz',    default=DEFAULT_CONTROL_FREQ_HZ,         type=int,           help='Control frequency in Hz (default: 48)', metavar='')
-    parser.add_argument('--aggregate',          default=DEFAULT_AGGREGATE,       type=str2bool,      help='Whether to aggregate physics steps (default: False)', metavar='')
     parser.add_argument('--duration_sec',       default=DEFAULT_DURATION_SEC,         type=int,           help='Duration of the simulation in seconds (default: 10)', metavar='')
     parser.add_argument('--output_folder',     default=DEFAULT_OUTPUT_FOLDER, type=str,           help='Folder where to save logs (default: "results")', metavar='')
     parser.add_argument('--colab',              default=DEFAULT_COLAB, type=bool,           help='Whether example is being run by a notebook (default: "False")', metavar='')
