--- gym_pybullet_drones/requirements.txt
@@ -1,69 +1,0 @@
-absl-py==1.4.0
-aiosignal==1.3.1
-attrs==22.2.0
-cachetools==5.2.1
-certifi==2022.12.7
-charset-normalizer==3.0.1
-click==8.0.4
-cloudpickle==2.2.1
-contourpy==1.0.7
-cycler==0.10.0
-distlib==0.3.6
-dm-tree==0.1.8
-filelock==3.9.0
-fonttools==4.38.0
-frozenlist==1.3.3
-google-auth==2.16.0
-google-auth-oauthlib==0.4.6
-grpcio==1.43.0
-seaborn==0.12.2
-gym==0.21.0
-gym-notices==0.0.8
-idna==3.4
-imageio==2.24.0
-importlib-metadata==4.13.0
-jsonschema==4.17.3
-kiwisolver==1.4.4
-lz4==4.3.2
-Markdown==3.4.1
-MarkupSafe==2.1.2
-matplotlib==3.6.3
-msgpack==1.0.4
-networkx==3.0
-numpy==1.24.1
-oauthlib==3.2.2
-packaging==23.0
-pandas==1.5.3
-Pillow==9.4.0
-platformdirs==2.6.2
-protobuf==3.20.1
-pyasn1==0.4.8
-pyasn1-modules==0.2.8
-pybullet==3.2.5
-pyparsing==3.0.9
-pyrsistent==0.19.3
-python-dateutil==2.8.2
-pytz==2022.7.1
-PyWavelets==1.4.1
-PyYAML==6.0
-ray==2.2.0
-requests==2.28.2
-requests-oauthlib==1.3.1
-rsa==4.9
-scikit-image==0.19.3
-scipy==1.10.0
-six==1.16.0
-stable-baselines3==1.5.0
-tabulate==0.9.0
-tensorboard==2.11.2
-tensorboard-data-server==0.6.1
-tensorboard-plugin-wit==1.8.1
-tensorboardX==2.5.1
-tifffile==2022.10.10
-torch==1.11.0
-typing_extensions==4.4.0
-urllib3==1.26.14
-virtualenv==20.17.1
-Werkzeug==2.2.2
-zipp==3.11.0
-wandb==0.13.8
